moving registrations and students functions to proper files
